pyproject.toml
[project]
name = "medical-data-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "katalria",email = "wawa0070@naver.com"}
]
readme = "README.md"
requires-python = "3.11.7"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "transformers (>=4.55.4,<5.0.0)",
    "streamlit (>=1.49.0,<2.0.0)",
    "langchain-openai (>=0.1.0,<1.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "langgraph (>=0.6.7,<0.7.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.optional-dependencies]
openai = ["langchain-openai (>=0.1.0,<1.0.0)"]
